api Integration/requirements.txt
jinja2==3.1.2
python-dotenv==1.0.0
secure-smtplib==0.1.1
pdfkit==1.0.0 